Remove multiple runtimes support from API
Having multiple runtime coexist is suboptimal from a performance
point of view and may even result in unnecessary dependencies on
the specific runtime used.
If extreme control is needed on how a service is executed in
addition to what is already provided by the single existing
runtime, it's advisable to spawn a new thread and do the
customization there.